feat: update sendFeatureDevTelemetry to include user context and opt in/out telemetry options
