+++ packages/core/src/amazonqFeatureDev/client/codewhispererruntime-2022-11-11.json
@@ -2210,6 +2210,12 @@
                 },
                 "product": {
                     "shape": "UserContextProductString"
+                },
+                "clientId": {
+                    "shape": "UUID"
+                },
+                "ideVersion": {
+                    "shape": "String"
                 }
             }
         },
